--- conf/ioc_types.json
@@ -1,13 +1,0 @@
-{
-  "ip": [
-    "sourceAddress",
-    "destinationAddress",
-    "deviceAddress"
-  ],
-  "domain": [
-    "destinationDomain"
-  ],
-  "md5": [
-    "fileHash"
-  ]
-}
