+++ tests/unit/stream_alert_cli/test_cli_config.py
@@ -21,7 +21,7 @@ from nose.tools import assert_equal, assert_true, assert_false
 from pyfakefs import fake_filesystem_unittest
 
 from stream_alert_cli.config import CLIConfig
-from tests.unit.helpers.base import basic_streamalert_config
+from tests.unit.helpers.config import basic_streamalert_config
 
 
 class TestCLIConfig(object):
@@ -40,6 +40,10 @@ class TestCLIConfig(object):
 
         self.fs_patcher.fs.create_file(
             '/conf/global.json', contents=json.dumps(config_data['global']))
+        self.fs_patcher.fs.create_file(
+            '/conf/threat_intel.json', contents=json.dumps(config_data['threat_intel']))
+        self.fs_patcher.fs.create_file(
+            '/conf/normalized_types.json', contents=json.dumps(config_data['normalized_types']))
         self.fs_patcher.fs.create_file(
             '/conf/lambda.json', contents=json.dumps(config_data['lambda']))
         self.fs_patcher.fs.create_file(
@@ -156,41 +160,57 @@ class TestCLIConfig(object):
 
         assert_equal(result, expected_result)
 
-    @patch('logging.Logger.info')
-    @patch('stream_alert_cli.config.CLIConfig.write')
-    def test_add_threat_intel_with_table_name(self, write_mock, log_mock):
+    def test_add_threat_intel_with_table_name(self):
         """CLI - Add Threat Intel config with default dynamodb table name"""
         threat_intel_info = {
-            'command': 'threat_intel',
+            'command': 'threat-intel',
             'debug': 'False',
-            'dynamodb_table': 'my_ioc_table',
-            'subcommand': 'enable'
+            'dynamodb_table_name': 'my_ioc_table',
+            'enable': True
         }
-
-        self.config.add_threat_intel(threat_intel_info)
-
-        expected_config = {'enabled': True, 'dynamodb_table': 'my_ioc_table'}
-
-        assert_equal(self.config['global']['threat_intel'], expected_config)
-        write_mock.assert_called()
-        log_mock.assert_called()
-
-    @patch('logging.Logger.info')
-    @patch('stream_alert_cli.config.CLIConfig.write')
-    def test_add_threat_intel_without_table_name(self, write_mock, log_mock):
-        """CLI - Add Threat Intel config without dynamodb table name from cli"""
-        threat_intel_info = {'command': 'threat_intel', 'debug': 'False', 'subcommand': 'enable'}
 
         self.config.add_threat_intel(threat_intel_info)
 
         expected_config = {
             'enabled': True,
-            'dynamodb_table': 'unit-testing_streamalert_threat_intel_downloader'
+            'dynamodb_table_name': 'my_ioc_table',
+            'excluded_iocs': {},
+            'normalized_ioc_types': {
+                'ip': [
+                    'sourceAddress',
+                    'destinationAddress'
+                ]
+            }
         }
 
-        assert_equal(self.config['global']['threat_intel'], expected_config)
-        write_mock.assert_called()
-        log_mock.assert_called()
+        assert_equal(self.config['threat_intel'], expected_config)
+
+    def test_add_threat_intel_without_table_name(self):
+        """CLI - Add Threat Intel config without dynamodb table name from cli"""
+        threat_intel_info = {
+            'command': 'threat-intel',
+            'debug': 'False',
+            'dynamodb_table_name': None,
+            'enable': True
+        }
+
+        del self.config['threat_intel']['dynamodb_table_name']
+
+        self.config.add_threat_intel(threat_intel_info)
+
+        expected_config = {
+            'enabled': True,
+            'dynamodb_table_name': 'unit-testing_streamalert_threat_intel_downloader',
+            'excluded_iocs': {},
+            'normalized_ioc_types': {
+                'ip': [
+                    'sourceAddress',
+                    'destinationAddress'
+                ]
+            }
+        }
+
+        assert_equal(self.config['threat_intel'], expected_config)
 
     @patch('logging.Logger.info')
     @patch('stream_alert_cli.config.CLIConfig.write')
